add uint8 render target
1. use render buffers instead of textured for deferred rendering
2. add the ability to change rendering ratio and use OpenGL for blitting

+++ fast_gauss/__init__.py
@@ -29,7 +29,7 @@ class GaussianRasterizer:
                  init_buffer_size: int = 32768,
                  init_texture_size: List[int] = [512, 512],
                  dtype=torch.float,
-                 tex_dtype=torch.half):
+                 tex_dtype=torch.uint8):
         super().__init__()
         self.raster_settings = raster_settings
